Use perl's own bool constants
It was perl 5.15.3 that began using the compiler bool type rather than
perl's own if it was available. Before that handy.h did not include
stdbool.h and so the "false" constant was not available. perl has long
provided TRUE and FALSE constants, so for best compatibility with older
perls (and weird systems without stdbool.h?) we should use those.